alert_api/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7